Do not close streams.
In C++, resources should generally not be closed explicitly. When an
instance goes out of scope, it is immediately destroyed, and its
destructor does the closing. Putting in explicit closes is usually a bad
idea, because it forces us to think "but what if an exception is
thrown?". Good code, as far as possible, relies upon destructors to do
resource cleanups.